+++ src/Provider.js
@@ -1,5 +1,5 @@
 import { Children, Component, createContext, createElement } from "react"
-import * as PropTypes from "./propTypes"
+import { shallowEqual } from "./utils/utils"
 
 const specialReactKeys = { children: true, key: true, ref: true }
 
@@ -35,23 +35,12 @@ class Provider extends Component {
         if (!nextProps) return null
         if (!prevState) return nextProps
 
-        // Maybe this warning is too aggressive?
-        if (
-            Object.keys(nextProps).filter(validStoreName).length !==
-            Object.keys(prevState).filter(validStoreName).length
-        )
-            console.warn(
+        const baseStores = { ...prevState, ...specialReactKeys } // mix in specialReactKeys, so that they are ignored in the diff
+        const newStores = { ...nextProps, ...specialReactKeys }
+        if (!shallowEqual(baseStores, newStores))
+            throw new Error(
                 "MobX Provider: The set of provided stores has changed. Please avoid changing stores as the change might not propagate to all children"
             )
-        if (!nextProps.suppressChangedStoreWarning)
-            for (let key in nextProps)
-                if (validStoreName(key) && prevState[key] !== nextProps[key])
-                    console.warn(
-                        "MobX Provider: Provided store '" +
-                            key +
-                            "' has changed. Please avoid replacing stores as the change might not propagate to all children"
-                    )
-
         return nextProps
     }
 }
@@ -64,5 +53,4 @@ function copyStores(from, to) {
 function validStoreName(key) {
     return !specialReactKeys[key] && key !== "suppressChangedStoreWarning"
 }
-
 export default Provider